Refactor.더미데이터 삭제, 프로토콜 파일분리
