+++ RideThis/Protocol/Bluetooth/Bluetooth.swift
@@ -1,0 +1,22 @@
+//
+//  Bluetooth.swift
+//  RideThis
+//
+//  Created by SeongKook on 8/30/24.
+//
+
+import Foundation
+
+
+protocol BluetoothManagerDelegate: AnyObject {
+    func didUpdateCadence(_ cadence: Double)
+    func didUpdateSpeed(_ speed: Double)
+    func didUpdateDistance(_ distance: Double)
+    func didUpdateCalories(_ calories: Double)
+    
+
+}
+
+protocol BluetoothViewDelegate: AnyObject {
+    func bluetoothDidTurnOff()
+}
