+++ RideThis/Protocol/CompetitionProtocol.swift
@@ -1,1 +1,5 @@
 import Foundation
+
+protocol CountViewControllerDelegate: AnyObject {
+    func countdownFinish()
+}
